# Optimize grep performance specifically for short queries