--- a/crates/maple/src/main.rs
+++ b/crates/maple/src/main.rs
@@ -1,6 +1,10 @@
 use clap::Parser;
 use cli::{Args, RunCmd};
 
+#[cfg(not(target_env = "msvc"))]
+#[global_allocator]
+static GLOBAL: tikv_jemallocator::Jemalloc = tikv_jemallocator::Jemalloc;
+
 const BUILD_TIME: &str = include!(concat!(env!("OUT_DIR"), "/compiled_at.txt"));
 
 mod built_info {