--- a/crates/maple_core/src/searcher/grep/mod.rs
+++ b/crates/maple_core/src/searcher/grep/mod.rs
@@ -9,6 +9,8 @@ use std::sync::Arc;
 use std::time::Instant;
 use tokio::sync::mpsc::unbounded_channel;
 
+use super::SearchInfo;
+
 #[derive(Debug)]
 pub struct SearchResult {
     pub matches: Vec<FileResult>,
@@ -21,14 +23,16 @@ pub async fn cli_search(paths: Vec<PathBuf>, matcher: Matcher) -> SearchResult {
 
     let stop_signal = Arc::new(AtomicBool::new(false));
 
-    let total_processed = Arc::new(AtomicUsize::new(0));
+    let search_info = SearchInfo {
+        total_processed: Arc::new(AtomicUsize::new(0)),
+    };
 
     {
-        let total_processed = total_processed.clone();
+        let search_info = search_info.clone();
         std::thread::Builder::new()
             .name("searcher-worker".into())
             .spawn(move || {
-                StoppableSearchImpl::new(paths, matcher, sender, stop_signal).run(total_processed)
+                StoppableSearchImpl::new(paths, matcher, sender, stop_signal).run(search_info)
             })
             .expect("Failed to spawn searcher worker thread");
     }
@@ -41,7 +45,9 @@ pub async fn cli_search(paths: Vec<PathBuf>, matcher: Matcher) -> SearchResult {
     while let Some(file_result) = receiver.recv().await {
         matches.push(file_result);
         total_matched += 1;
-        let total_processed = total_processed.load(std::sync::atomic::Ordering::Relaxed);
+        let total_processed = search_info
+            .total_processed
+            .load(std::sync::atomic::Ordering::Relaxed);
 
         if total_matched % 16 == 0 || total_processed % 16 == 0 {
             let now = Instant::now();
@@ -52,7 +58,9 @@ pub async fn cli_search(paths: Vec<PathBuf>, matcher: Matcher) -> SearchResult {
         }
     }
 
-    let total_processed = total_processed.load(std::sync::atomic::Ordering::SeqCst) as u64;
+    let total_processed = search_info
+        .total_processed
+        .load(std::sync::atomic::Ordering::SeqCst) as u64;
 
     SearchResult {
         matches,