--- a/crates/maple_core/src/searcher/mod.rs
+++ b/crates/maple_core/src/searcher/mod.rs
@@ -6,16 +6,30 @@ pub mod tagfiles;
 use crate::stdio_server::Vim;
 use icon::Icon;
 use ignore::{WalkBuilder, WalkParallel};
+use parking_lot::RwLock;
 use paths::AbsPathBuf;
 use serde::{Deserialize, Serialize};
 use std::path::PathBuf;
 use std::sync::atomic::{AtomicBool, AtomicUsize};
 use std::sync::Arc;
+use types::Rank;
 
 // Shareable searching info.
 #[derive(Clone, Debug)]
 pub struct SearchInfo {
+    pub lowest_rank: Arc<RwLock<Rank>>,
+    pub total_matched: Arc<AtomicUsize>,
     pub total_processed: Arc<AtomicUsize>,
+}
+
+impl SearchInfo {
+    pub fn new() -> Self {
+        Self {
+            lowest_rank: Arc::new(RwLock::new(Rank::default())),
+            total_matched: Arc::new(AtomicUsize::new(0)),
+            total_processed: Arc::new(AtomicUsize::new(0)),
+        }
+    }
 }
 
 #[derive(Debug, Clone)]