--- a/crates/matcher/src/algo/mod.rs
+++ b/crates/matcher/src/algo/mod.rs
@@ -1,4 +1,6 @@
+pub mod fzf;
 pub mod fzy;
+pub mod nucleo;
 pub mod skim;
 pub mod substring;
 
@@ -8,9 +10,11 @@ use types::{CaseMatching, FuzzyText};
 // TODO: Integrate https://github.com/nomad/norm for fzf algo.
 #[derive(Debug, Clone, Copy, Default)]
 pub enum FuzzyAlgorithm {
-    Skim,
     #[default]
     Fzy,
+    Skim,
+    FzfV2,
+    Nucleo,
 }
 
 impl std::str::FromStr for FuzzyAlgorithm {
@@ -25,6 +29,8 @@ impl<T: AsRef<str>> From<T> for FuzzyAlgorithm {
         match algo.as_ref().to_lowercase().as_str() {
             "skim" => Self::Skim,
             "fzy" => Self::Fzy,
+            "fzf-v2" => Self::FzfV2,
+            "nucleo" => Self::Nucleo,
             _ => Self::Fzy,
         }
     }
@@ -45,6 +51,8 @@ impl FuzzyAlgorithm {
         let fuzzy_result = match self {
             Self::Fzy => fzy::fuzzy_indices(text, query, case_matching),
             Self::Skim => skim::fuzzy_indices(text, query, case_matching),
+            Self::FzfV2 => fzf::fuzzy_indices_v2(text, query),
+            Self::Nucleo => nucleo::fuzzy_indices(text, query, case_matching),
         };
         fuzzy_result.map(|MatchResult { score, indices }| {
             let mut indices = indices;