When the query is merely one or two chars, the tokio mpsc channel could be heavily used which can be one bottleneck as we observed before. This PR optimizes for this scenario, ~20% boost gained, the impact of this tactic on the performance of other cases is negligible.

```
hyperfine --warmup 3 -L query c './maple-master grep --lib-ripgrep --cmd-dir /home/xlc/Data2/github.com/paritytech/polkadot-sdk/ {query}' './target/release/maple grep --lib-ripgrep --cmd-dir /home/xlc/Data2/github.com/paritytech/polkadot-sdk/ {query}'
Benchmark 1: ./maple-master grep --lib-ripgrep --cmd-dir /home/xlc/Data2/github.com/paritytech/polkadot-sdk/ c
  Time (mean ± σ):     616.3 ms ±  12.6 ms    [User: 1011.6 ms, System: 530.5 ms]
  Range (min … max):   590.6 ms … 629.2 ms    10 runs
 
Benchmark 2: ./target/release/maple grep --lib-ripgrep --cmd-dir /home/xlc/Data2/github.com/paritytech/polkadot-sdk/ c
  Time (mean ± σ):     509.0 ms ±  16.5 ms    [User: 850.1 ms, System: 376.8 ms]
  Range (min … max):   465.7 ms … 525.2 ms    10 runs
 
Summary
  ./target/release/maple grep --lib-ripgrep --cmd-dir /home/xlc/Data2/github.com/paritytech/polkadot-sdk/ c ran
    1.21 ± 0.05 times faster than ./maple-master grep --lib-ripgrep --cmd-dir /home/xlc/Data2/github.com/paritytech/polkadot-sdk/ c
```

```
hyperfine --warmup 3 -L query client './maple-master grep --lib-ripgrep --cmd-dir /home/xlc/Data2/github.com/paritytech/polkadot-sdk/ {query}' './target/release/maple grep --lib-ripgrep --cmd-dir /home/xlc/Data2/github.com/paritytech/polkadot-sdk/ {query}'
Benchmark 1: ./maple-master grep --lib-ripgrep --cmd-dir /home/xlc/Data2/github.com/paritytech/polkadot-sdk/ client
  Time (mean ± σ):     329.5 ms ±   8.2 ms    [User: 514.3 ms, System: 233.7 ms]
  Range (min … max):   315.1 ms … 342.6 ms    10 runs
 
Benchmark 2: ./target/release/maple grep --lib-ripgrep --cmd-dir /home/xlc/Data2/github.com/paritytech/polkadot-sdk/ client
  Time (mean ± σ):     331.6 ms ±  10.5 ms    [User: 529.8 ms, System: 217.8 ms]
  Range (min … max):   311.0 ms … 350.0 ms    10 runs
 
Summary
  ./maple-master grep --lib-ripgrep --cmd-dir /home/xlc/Data2/github.com/paritytech/polkadot-sdk/ client ran
    1.01 ± 0.04 times faster than ./target/release/maple grep --lib-ripgrep --cmd-dir /home/xlc/Data2/github.com/paritytech/polkadot-sdk/ client
 ```