@liuchengxu 2024-04-21
Switching to Jemalloc gives us at least 5 ~ 10% boost, for the short queries, the number can be up to 20%. Therefore, the final boost in this PR is 5 ~ 40%, depending on the length of query. The shorter the query, the greater the boost in performance.
```
hyperfine --warmup 3 -L query c './maple-master grep --lib-ripgrep --cmd-dir /home/xlc/Data2/github.com/paritytech/polkadot-sdk/ {query}' './target/release/maple grep --lib-ripgrep --cmd-dir /home/xlc/Data2/github.com/paritytech/polkadot-sdk/ {query}'
Benchmark 1: ./maple-master grep --lib-ripgrep --cmd-dir /home/xlc/Data2/github.com/paritytech/polkadot-sdk/ c
  Time (mean ± σ):     615.2 ms ±  48.9 ms    [User: 987.7 ms, System: 545.4 ms]
  Range (min … max):   527.4 ms … 674.5 ms    10 runs
 
Benchmark 2: ./target/release/maple grep --lib-ripgrep --cmd-dir /home/xlc/Data2/github.com/paritytech/polkadot-sdk/ c
  Time (mean ± σ):     438.4 ms ±   8.9 ms    [User: 762.5 ms, System: 312.9 ms]
  Range (min … max):   423.5 ms … 450.5 ms    10 runs
 
Summary
  ./target/release/maple grep --lib-ripgrep --cmd-dir /home/xlc/Data2/github.com/paritytech/polkadot-sdk/ c ran
    1.40 ± 0.12 times faster than ./maple-master grep --lib-ripgrep --cmd-dir /home/xlc/Data2/github.com/paritytech/polkadot-sdk/ c
```

```
hyperfine --warmup 3 -L query client './maple-master grep --lib-ripgrep --cmd-dir /home/xlc/Data2/github.com/paritytech/polkadot-sdk/ {query}' './target/release/maple grep --lib-ripgrep --cmd-dir /home/xlc/Data2/github.com/paritytech/polkadot-sdk/ {query}'
Benchmark 1: ./maple-master grep --lib-ripgrep --cmd-dir /home/xlc/Data2/github.com/paritytech/polkadot-sdk/ client
  Time (mean ± σ):     313.3 ms ±  14.6 ms    [User: 490.7 ms, System: 222.9 ms]
  Range (min … max):   287.1 ms … 329.8 ms    10 runs
 
Benchmark 2: ./target/release/maple grep --lib-ripgrep --cmd-dir /home/xlc/Data2/github.com/paritytech/polkadot-sdk/ client
  Time (mean ± σ):     292.8 ms ±  12.7 ms    [User: 480.8 ms, System: 189.1 ms]
  Range (min … max):   260.4 ms … 301.9 ms    10 runs
 
Summary
  ./target/release/maple grep --lib-ripgrep --cmd-dir /home/xlc/Data2/github.com/paritytech/polkadot-sdk/ client ran
    1.07 ± 0.07 times faster than ./maple-master grep --lib-ripgrep --cmd-dir /home/xlc/Data2/github.com/paritytech/polkadot-sdk/ client
```